fix: fixes circular imports in utils
