+++ mobile/src/screens/WelcomeScreen.tsx
@@ -2,7 +2,8 @@ import React, { useEffect } from 'react';
 import { View, StyleSheet, Text } from 'react-native';
 import { LoginButton } from '../components/buttons/AuthButtons';
 import { useNavigation } from '@react-navigation/native';
-import { hasOnboarded, hasValidCredentials } from '../utils/auth';
+import { hasValidCredentials } from '../utils/auth';
+import { hasOnboarded } from '../utils/db';
 import { SvgUri } from 'react-native-svg';
 import { colors } from '../../assets/colors';
 import { useAuth0 } from 'react-native-auth0';
