+++ mobile/src/utils/db.tsx
@@ -662,3 +662,19 @@ export const fetchEvents = async () => {
     console.error('Error:', error);
   }
 };
+
+/**
+ * Checks if the user has completed the onboarding process.
+ * @returns A promise that resolves to a boolean indicating whether the user has completed onboarding.
+ */
+export const hasOnboarded = async () => {
+  try {
+    const userData = await fetchUserData();
+    if (userData) {
+      return true;
+    }
+    return false;
+  } catch (error) {
+    console.error('Error:', error);
+  }
+};
